[fix] update alert documentation (#10)
- document Alert component
- document Alert subcomponents (Header, Body, Link)
