--- src/components/Alert/Alert.stories.mdx
@@ -1,68 +1,0 @@
-import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-
-import Alert from './Alert';
-
-<Meta
-  title="Components"
-  component={Alert}
-  subcomponents={{ 'Alert.Header': Alert.Header, 'Alert.Body': Alert.Body }}
-/>
-
-export const Template = (args) => (
-  <Alert {...args}>This is a sample alert.</Alert>
-);
-
-# Alert
-
-Use to provide contextual feedback messages for typical user actions. Choose from a handful of available and flexible alert messages.
-
-## Available Props
-
-### Sample:
-
-<Story name="Alert">{Template.bind({})}</Story>
-
-<ArgsTable
-  component={{ 'Alert.Header': Alert.Header, 'Alert.Body': Alert.Body }}
-/>
-
-## Button styling variants
-
-<Canvas>
-  <div>
-    <Alert variant="success" message="Default">
-      <Alert.Header>Title</Alert.Header>
-      <Alert.Body>
-        Success content here: <Alert.Link href="#">Test Link</Alert.Link>
-      </Alert.Body>
-    </Alert>
-    <Alert variant="danger" message="Default">
-      <Alert.Header>Title</Alert.Header>
-      <Alert.Body>
-        Danger content here: <Alert.Link href="#">Test Link</Alert.Link>
-      </Alert.Body>
-    </Alert>
-    <Alert variant="warning" message="Default">
-      <Alert.Header>Title</Alert.Header>
-      <Alert.Body>
-        Warning content here: <Alert.Link href="#">Test Link</Alert.Link>
-      </Alert.Body>
-    </Alert>
-    <Alert variant="info" message="Default">
-      <Alert.Header>Title</Alert.Header>
-      <Alert.Body>
-        Info content here: <Alert.Link href="#">Test Link</Alert.Link>
-      </Alert.Body>
-    </Alert>
-  </div>
-</Canvas>
-
-## Button sizing variants
-
-## Disabled button
-
-## Active state
-
-## Button tags
-
-## Button blocks
