+++ src/components/Alert/stories/Alert.intro.stories.mdx
@@ -1,0 +1,32 @@
+import { ArgsTable, Canvas, Meta } from '@storybook/addon-docs';
+
+import { Alert } from '../..';
+
+<Meta
+  title="Components/Alert/Intro"
+  component={Alert}
+  subcomponents={{
+    'Alert.Header': Alert.Header,
+    'Alert.Body': Alert.Body,
+    'Alert.Link': Alert.Link,
+  }}
+/>
+
+# Alert
+
+Use to provide contextual feedback messages for typical user actions. Choose from a handful of available and flexible alert messages.
+
+## Example
+
+<Canvas>
+  <Alert>
+    <Alert.Header>Sample Alert</Alert.Header>
+    <Alert.Body>
+      sample content here: <Alert.Link href="#">Test Link</Alert.Link>
+    </Alert.Body>
+  </Alert>
+</Canvas>
+
+## Available Props
+
+<ArgsTable />
